[syl-dark-styles] Replace AllVAriants with statesheet for action styles
