+++ __docs__/wonder-blocks-styles/action-styles.stories.tsx
@@ -38,6 +38,12 @@ export default {
                 version={packageConfig.version}
             />
         ),
+        parameters: {
+            chromatic: {
+                // Disabling because this is already covered by the statesheet
+                disableSnapshot: true,
+            },
+        },
     },
 } as Meta<any>;
 
@@ -137,13 +143,6 @@ export const InverseOutline: Story = {
     globals: {
         backgrounds: {
             value: "neutralStrong",
-        },
-    },
-    parameters: {
-        chromatic: {
-            // Disabling because this is already covered by the All variants
-            // stories.
-            disableSnapshot: true,
         },
     },
 };
